Update Enable-MailboxAuditing.ps1
A customer of mine is running this script for 86k mailboxes and it is running for several days. Customer has no progress indication. That commit has a in lab tested suggestion of a progress bar to make it easier for customers and PFEs to track the progress in big environments.