Filter out deleted users
When using the $filter=signInActivity/lastSignInDateTime filter on the users endpoint, it will return deleted users too. Those will have a lastsignindatetime but all other properties will be empty. It creates confusion when looking at the report.
I suggest we simply filter out result without a UPN (indicating it is a deleted user).